storage: extract parts of region_info
