+++ components/tikv_misc/src/lib.rs
@@ -23,6 +23,7 @@ pub mod cop_table_consts;
 pub mod cop_props;
 pub mod flow_stats;
 pub mod keys;
+pub mod kv_region_info;
 pub mod mvcc_lock;
 pub mod mvcc_types;
 pub mod mvcc_write;
